Rename ``views`` module to ``view`` and any references in the project - Update imports and docstrings for ``view`` in modules and tests - Update Sphinx docs
