+++ tests/unit/db/design_document_tests.py
@@ -26,7 +26,7 @@ import unittest
 
 from cloudant.document import Document 
 from cloudant.design_document import DesignDocument
-from cloudant.views import View, QueryIndexView
+from cloudant.view import View, QueryIndexView
 from cloudant.errors import CloudantArgumentError, CloudantException
 
 from .unit_t_db_base import UnitTestDbBase
